Moved logic for at_time to codegen_c_visitor
