+++ src/codegen/codegen_c_visitor.cpp
@@ -15,6 +15,7 @@
 #include "codegen/codegen_helper_visitor.hpp"
 #include "codegen/codegen_naming.hpp"
 #include "config/config.h"
+#include "lexer/token_mapping.hpp"
 #include "parser/c11_driver.hpp"
 #include "utils/logger.hpp"
 #include "utils/string_utils.hpp"
@@ -1331,6 +1332,10 @@ void CodegenCVisitor::print_function_call(FunctionCall& node) {
         printer->add_text(internal_method_arguments());
         if (!arguments.empty()) {
             printer->add_text(", ");
+        }
+    } else if (nmodl::details::needs_neuron_thread_first_arg(function_name)) {
+        if (arguments.front()->get_node_name() != "nt") {
+            arguments.insert(arguments.begin(), std::make_shared<ast::String>("nt"));
         }
     }
