+++ src/lexer/token_mapping.hpp
@@ -25,4 +25,8 @@ parser::NmodlParser::token_type token_type(const std::string& name);
 std::vector<std::string> get_external_variables();
 std::vector<std::string> get_external_functions();
 
+namespace details {
+bool needs_neuron_thread_first_arg(const std::string& token);
+}  // namespace details
+
 }  // namespace nmodl
